+++ CHANGELOG.rst
@@ -35,6 +35,10 @@ As a consequence of this change:
 
 Thanks :user:`ddelange` for the PR.
 
+Previously-deprecated API have been removed, including:
+
+- `default` and `missing` parameters, which were replaced by `dump_default` and `load_default` in 3.13.0 (:pr:`1742`).
+
 3.23.2 (2024-12-18)
 *******************